// src/app/models/set.interface.ts
export interface Set {
  object: string;
  id: string;
  code: string;
  tcgplayer_id: number;
  name: string;
  uri: string;
  scryfall_uri: string;
  search_uri: string;
  released_at: string;
  set_type: string;
  card_count: number;
  digital: boolean;
  nonfoil_only: boolean;
  foil_only: number;
  icon_svg_uri: string;
}